agents/agentkit/Python/requirements.txt
web3==6.15.1
python-dotenv==1.0.0
coinbase==2.1.0
eth-utils==2.3.1
eth-typing==3.5.2